Fixed 0.2.2

+++ src/main/java/minegame159/meteorclient/gui/renderer/TriangleOperation.java
@@ -28,16 +28,16 @@ public class TriangleOperation extends Operation {
 
         double x = ((this.x - oX) * cos) - ((this.y - oY) * sin) + oX;
         double y = ((this.y - oY) * cos) + ((this.x - oX) * sin) + oY;
-        renderer.quadBuf.vertex(x, y, 0).texture(GuiRenderer.TEX_QUAD.x, GuiRenderer.TEX_QUAD.y).color(color.r, color.g, color.b, color.a).next();
+        renderer.quadBuf.vertex(x, y, 0).texture((float) GuiRenderer.TEX_QUAD.x, (float) GuiRenderer.TEX_QUAD.y).color(color.r, color.g, color.b, color.a).next();
 
         x = ((this.x + size - oX) * cos) - ((this.y - oY) * sin) + oX;
         y = ((this.y - oY) * cos) + ((this.x + size - oX) * sin) + oY;
-        renderer.quadBuf.vertex(x, y, 0).texture(GuiRenderer.TEX_QUAD.x + GuiRenderer.TEX_QUAD.width, GuiRenderer.TEX_QUAD.y).color(color.r, color.g, color.b, color.a).next();
+        renderer.quadBuf.vertex(x, y, 0).texture((float) (GuiRenderer.TEX_QUAD.x + GuiRenderer.TEX_QUAD.width), (float) GuiRenderer.TEX_QUAD.y).color(color.r, color.g, color.b, color.a).next();
 
         x = ((this.x + size / 2 - oX) * cos) - ((this.y + size / 2 - oY) * sin) + oX;
         y = ((this.y + size / 2 - oY) * cos) + ((this.x + size / 2 - oX) * sin) + oY;
-        renderer.quadBuf.vertex(x, y, 0).texture(GuiRenderer.TEX_QUAD.x + GuiRenderer.TEX_QUAD.width, GuiRenderer.TEX_QUAD.y + GuiRenderer.TEX_QUAD.height).color(color.r, color.g, color.b, color.a).next();
-        renderer.quadBuf.vertex(x, y, 0).texture(GuiRenderer.TEX_QUAD.x, GuiRenderer.TEX_QUAD.y + GuiRenderer.TEX_QUAD.height).color(color.r, color.g, color.b, color.a).next();
+        renderer.quadBuf.vertex(x, y, 0).texture((float) (GuiRenderer.TEX_QUAD.x + GuiRenderer.TEX_QUAD.width), (float) (GuiRenderer.TEX_QUAD.y + GuiRenderer.TEX_QUAD.height)).color(color.r, color.g, color.b, color.a).next();
+        renderer.quadBuf.vertex(x, y, 0).texture((float) GuiRenderer.TEX_QUAD.x, (float) (GuiRenderer.TEX_QUAD.y + GuiRenderer.TEX_QUAD.height)).color(color.r, color.g, color.b, color.a).next();
     }
 
     @Override
